Move Job payment validation related code into separate file
