+++ blockchain/interceptors.go
@@ -55,51 +55,6 @@ type paymentHandlerType interface {
 	completePayment(error) error
 }
 
-type jobPaymentHandler struct {
-	p                 Processor
-	md                metadata.MD
-	jobAddressBytes   []byte
-	jobSignatureBytes []byte
-}
-
-func newJobPaymentHandler(p Processor, md metadata.MD) *jobPaymentHandler {
-	return &jobPaymentHandler{p: p, md: md}
-}
-
-func (h *jobPaymentHandler) validatePayment() error {
-	jobAddressMd, ok := h.md[JobAddressHeader]
-	if !ok {
-		return status.Errorf(codes.InvalidArgument, "missing snet-job-address")
-	}
-
-	h.jobAddressBytes = common.FromHex(jobAddressMd[0])
-
-	jobSignatureMd, ok := h.md[JobSignatureHeader]
-	if !ok {
-		return status.Errorf(codes.InvalidArgument, "missing snet-job-signature")
-	}
-
-	h.jobSignatureBytes = common.FromHex(jobSignatureMd[0])
-	valid := h.p.IsValidJobInvocation(h.jobAddressBytes, h.jobSignatureBytes)
-	if !valid {
-		return status.Errorf(codes.Unauthenticated, "job invocation not valid")
-	}
-
-	return nil
-}
-
-func (h *jobPaymentHandler) completePayment(err error) error {
-	if err == nil {
-		h.p.CompleteJob(h.jobAddressBytes, h.jobSignatureBytes)
-	}
-	return err
-}
-
-func noOpInterceptor(srv interface{}, ss grpc.ServerStream, info *grpc.StreamServerInfo,
-	handler grpc.StreamHandler) error {
-	return handler(srv, ss)
-}
-
 func getBigInt(md metadata.MD, key string) (*big.Int, error) {
 	array := md.Get(key)
 	if len(array) == 0 {
@@ -121,4 +76,9 @@ func getBytes(md metadata.MD, key string) ([]byte, error) {
 		return nil, status.Errorf(codes.InvalidArgument, "missing \"%v\"", key)
 	}
 	return common.FromHex(value[0]), nil
+}
+
+func noOpInterceptor(srv interface{}, ss grpc.ServerStream, info *grpc.StreamServerInfo,
+	handler grpc.StreamHandler) error {
+	return handler(srv, ss)
 }
